移除旧的图表库
