+++ gradle/libs.versions.toml
@@ -1,6 +1,5 @@
 [versions]
 androidGradlePlugin = "8.7.3"
-charty = "2.1.0-beta03.2"
 vico = "2.1.3"
 kotlin = "2.1.20"
 kotlinComposeCompilerPlugin = "2.1.20"
@@ -30,7 +29,6 @@ swipe = "1.3.0"
 serializationJson = "1.7.3"
 
 [libraries]
-charty = { module = "com.himanshoe:charty", version.ref = "charty" }
 vico-compose-m3 = { group = "com.patrykandpatrick.vico", name = "compose-m3", version.ref = "vico" }
 hilt-android = { module = "com.google.dagger:hilt-android", version.ref = "hilt" }
 hilt-android-compiler = { module = "com.google.dagger:hilt-android-compiler", version.ref = "hilt" }
